+++ src/cpp/abaqus_tools.h
@@ -144,10 +144,10 @@ namespace abaqusTools{
     }
 
     template< typename T >
-    std::vector< T > contractAbaqusStandardStressVector( const std::vector< T > &full_abaqus_vector,
-                                                         const int &NDI, const int &NSHR ){
+    std::vector< T > contractAbaqusNTENSVector( const std::vector< T > &full_abaqus_vector,
+                                                 const int &NDI, const int &NSHR ){
         /*!
-         * Contract stress and strain type components from full Abaqus vectors
+         * Contract stress and strain type components from full Abaqus vectors.
          *
          * See the Abaqus documentation > Introduction & Spatial Modeling > Conventions chapter > Convention used for stress
          * and strain components.
@@ -160,10 +160,20 @@ namespace abaqusTools{
          *
          * \f$ \left ( \epsilon_{11}, \epsilon_{22}, \epsilon_{33}, \gamma_{12}, \gamma_{13}, \gamma_{23} \right ) \f$
          *
+         * The stress vector components for Abaqus/Explicit (VUMAT) are
+         *
+         * \f$ \left ( \sigma_{11}, \sigma_{22}, \sigma_{33}, \tau_{12}, \tau_{23}, \tau_{13} \right ) \f$
+         *
+         * and the strain vector components match as
+         *
+         * \f$ \left ( \epsilon_{11}, \epsilon_{22}, \epsilon_{33}, \gamma_{12}, \gamma_{23}, \gamma_{13} \right ) \f$
+         *
          * where components that are zero-valued by definition, e.g. plane stress, are omitted. The shear strain is the
          * engineering shear strain where
          *
          * \f$ \gamma_{ij} = \epsilon_{ij} + \epsilon_{ji} \f$
+         *
+         * for both Abaqus/Standard and Abaqus/Explicit.
          *
          * \param &full_abaqus_vector: a previously expanded abaqus stress-type vector. Length 6.
          * \param &NDI: The number of direct components.